Project: GUI Framework
Hi, I'm currently experimenting with writing a GUI Abstraction framework (meaning I want the OS to render buttons and stuff) and
I made some nice progress with how I want the API to look and feel and how to make it work in rust. Yesterday I finally started on how to deal with user events (e.g. mouse clicks).
My background is actually C/C++ but I recently did a few Typescript/AngularJS Frontend Projects. I really liked how Angular uses [https://rxjs.dev/guide/observable](Observables) for dealing with events and that's why I would really like to do something similar in Rust. I'm currently looking into and playing around with Streams and it looks promising.

I tried to keep the project described in the PR a generic GUI frameworks with the only assumption that it uses Streams for the user event handling.

I did add some example code on how such an API could be used. It's meant to give an idea and is not meant to compile. I hope this makes things a bit more clear.

## Character
I do believe a new character might be useful. Someone with a background in Javascript or similar with a focus on **Frontend-Applications**. I feel the current character are more Backend/Low-Level and not Front-End Developers. I know this is not a common scenario with Rust missing a UI Framework but that hopefully changes in the future. Let me know and I can write something up.

## Note:
I probably mixed up some stuff. Some parts are written from the view of someone *Developing a UI framework* in rust while other aspects come from someone *using a UI framework* in rust.
I'm not sure if it's worth splitting those two up.