Restructure Polish goal and initiative pages
This change makes it so the Polish goal page and its entry in the
roadmap have the same set of initiatives and in the same order. It also
creates a basic landing page for each initiative that was missing, and
includes a list of milestones for each initiative. Future commits will
fill in more of the landing pages.